[sagecell] Fix install()
sagecell/config/argparse.txt: change "_parser_open" description;
rename "_parser_show" to "_parser_open";
rename "_parser_run" to "_parser_start".
sagecell/sagecell.py fix install();
rename show() to open_sagemathcell();
rename run() to start().

+++ sagecell/sagecell.py
@@ -2,7 +2,7 @@
 
 from argparse import ArgumentParser
 from os import remove
-from os.path import dirname, expanduser, join
+from os.path import dirname, exists, expanduser, isfile, join
 from subprocess import check_output, CalledProcessError
 from sys import argv, exit
 
@@ -49,7 +49,7 @@ def install():
             (answer_lower == 'nix')):
         print(messages["_error_Internet"])
         exit(0)
-    # Check git version/
+    # Check git version
     try:
         git_version = check_output("git --version", shell=True)
     except CalledProcessError:
@@ -64,8 +64,7 @@ def install():
         with settings(prompts={'Do you want to continue [Y/n]? ': 'Y'}):
             local("sudo apt-get update")
             local("sudo apt-get install git")
-    # /Check git version
-    # Check pip/
+    # Check pip
     try:
         pip_version = check_output("pip -V", shell=True)
     except CalledProcessError:
@@ -73,10 +72,12 @@ def install():
         local("cd %s; wget https://bootstrap.pypa.io/get-pip.py" %
                 downloads_path)
         local("cd %s; sudo python get-pip.py" % downloads_path)
-        remove(join(downloads_path, "get-pip.py"))
-    # /Check pip
+        pip_path = join(downloads_path, "get-pip.py")
+        if exists(pip_path) and isfile(pip_path):
+            remove(pip_path)
     # Install npm
     with settings(prompts={'Do you want to continue [Y/n]? ': 'Y'}):
+        local("sudo apt-get update")
         local("sudo apt-get install npm")
     # Make an alias (-g -- globally)
     local("sudo ln -s /usr/bin/nodejs /usr/bin/node")
@@ -97,11 +98,12 @@ def install():
     local("mv %s %s" % (join(sc_build_path, "github/sage"), sc_build_path))
     # Install Sage dependencies
     with settings(prompts={'Do you want to continue [Y/n]? ': 'Y'}):
+        local("sudo apt-get update")
         local("sudo apt-get install gcc m4 make perl tar")
     # Build Sage
     sage_path = join(sc_build_path, "sage")
     local("cd %s; make start" % sage_path)
-    # We need IPython stuff, not present in spkg/
+    # We need IPython stuff, not present in spkg
     local("cd %s; rm -rf IPython*" % join(sage_path,
             "local/lib/python/site-packages"))
     local("cd %s; rm -rf ipython*" % join(sage_path,
@@ -109,12 +111,10 @@ def install():
     ipython_path = join(sc_build_path, "github/ipython")
     local("mv %s %s" % (ipython_path, sage_path))
     local("cd %s; ../sage setup.py develop" % join(sage_path, "ipython"))
-    # /We need IPython stuff not present in spkg
-    # We need a cutting-edge matplotlib/
+    # We need a cutting-edge matplotlib
     matplotlib_path = join(sc_build_path, "github/matplotlib")
     local("mv %s %s" % (matplotlib_path, sage_path))
     local("cd %s; ../sage setup.py install" % join(sage_path, "matplotlib"))
-    # /We need a cutting-edge matplotlib
     # Install ecdsa, lockfile, paramiko, sockjs-tornado
     local("cd %s; sudo ./sage -pip install --no-deps --upgrade "
           "ecdsa" % sage_path)
@@ -128,35 +128,33 @@ def install():
     sagecell_path_old = join(sc_build_path, "github/sagecell")
     local("mv %s %s" % (sagecell_path_old, sage_path))
     local("cd %s; ../sage -sh -c \"make -B\"" % join(sage_path, "sagecell"))
-    # Configuration/
+    # Configuration
     local("cd %s; cp config_default.py config.py" % join(sage_path,
                                                          "sagecell"))
-    # Change config.py file/
-    sagecell_path_new = join(sage_path, "sagecell")
-    config_path = join(sagecell_path_new, "config.py")
-    sage_script_path = join(sage_path, "sage")
-    with open(config_path, 'r') as f:
-        lines = f.readlines()
-    for i, line in enumerate(lines):
-        if line == "sage = \"\"\n":
-            lines[i] = "sage = \"%s\"\n" % sage_script_path
-            break
-    with open(config_path, 'w') as f:
-        for line in lines:
-            f.write(line)
-    # /Change config.py file
-    # /Configuration
-    # Install python-dev for psutil installation
-    with settings(prompts={'Do you want to continue [Y/n]? ': 'Y'}):
-        local("sudo apt-get install python-dev")
-    # Install psutil, SQLAlchemy
-    local("sudo pip install psutil SQLAlchemy")
-    # Install IPython Notebook
-    local("sudo pip install 'ipython[notebook]'")
-    # Install sockjs.tornado
-    local("sudo pip install sockjs-tornado")
-    # Install python-daemon
-    local("sudo pip install python-daemon")
+    # Check psutil
+    psutil_dest_path = expanduser("~/sc_build/sage/local/lib/python2.7/psutil")
+    if not exists(psutil_dest_path):
+        psutil_source_path = "/usr/local/lib/python2.7/dist-packages/psutil"
+        if not exists(psutil_source_path):
+            with settings(prompts={'Do you want to continue [Y/n]? ': 'Y'}):
+                local("sudo apt-get update")
+                # Install python-dev for psutil installation
+                local("sudo apt-get install python-dev")
+            # Install psutil
+            local("sudo pip install psutil")
+        # Copy psutil
+        local("cp %s %s" % (psutil_source_path, psutil_dest_path))
+    # Check SQLAlchemy
+    sqlalchemy_dest_path = expanduser("~/sc_build/sage/local/lib/python2.7/"
+                                      "sqlalchemy")
+    if not exists(sqlalchemy_dest_path):
+        sqlalchemy_source_path = ("/usr/local/lib/python2.7/dist-packages/"
+                                  "sqlalchemy")
+        if not exists(sqlalchemy_source_path):
+            # Install SQLAlchemy
+            local("sudo pip install SQLAlchemy")
+        # Copy SQLAlchemy
+        local("cp %s %s" % (sqlalchemy_source_path, sqlalchemy_dest_path))
 
 def main():
     """Main function"""
@@ -180,16 +178,16 @@ def parse_command_line_args():
             description=argparse["_parser_install"],
             help=argparse["_parser_install"])
     parser_install.set_defaults(function_name=install)
-    # Create the parser for the "run" subcommand
-    parser_list = subparsers.add_parser("run",
-            description=argparse["_parser_run"],
-            help=argparse["_parser_run"])
-    parser_list.set_defaults(function_name=run)
-    # Create the parser for the "show" subcommand
-    parser_list = subparsers.add_parser("show",
-            description=argparse["_parser_show"],
-            help=argparse["_parser_show"])
-    parser_list.set_defaults(function_name=show)
+    # Create the parser for the "start" subcommand
+    parser_list = subparsers.add_parser("start",
+            description=argparse["_parser_start"],
+            help=argparse["_parser_start"])
+    parser_list.set_defaults(function_name=start)
+    # Create the parser for the "open" subcommand
+    parser_list = subparsers.add_parser("open",
+            description=argparse["_parser_open"],
+            help=argparse["_parser_open"])
+    parser_list.set_defaults(function_name=open_sagemathcell)
     # Create the parser for the "ssh" subcommand
     parser_list = subparsers.add_parser("ssh",
             description=argparse["_parser_ssh"],
@@ -200,23 +198,17 @@ def parse_command_line_args():
         exit(0) # Clean exit without any errors/problems
     return parser.parse_args()
 
-def run():
-    """Run the Sage Cell Server"""
+def open_sagemathcell():
+    """Open browser with the SageMathCell"""
 
-    sagecell_path = expanduser("~/sc_build/sage/sagecell")
-    print("Shut down the Sage Cell Server: Ctrl+c")
-    local("cd %s; ./web_server.py" % sagecell_path)
-
-def show():
-    """Open browser with the Sage Cell Server"""
-
-    local("xdg-open http://localhost:8888") # TODO: check port from config.py
+    local("xdg-open http://localhost:8888")
 
 def ssh():
     """Setup SSH for auto login to localhost without a password"""
 
-    # Install  openssh-server
+    # Install openssh-server
     with settings(prompts={'Do you want to continue [Y/n]? ': 'Y'}):
+        local("sudo apt-get update")
         local("sudo apt-get install openssh-server")
     # Create a public and a private keys using the ssh-keygen command
     local("ssh-keygen -t rsa -b 4096 -N '' -f ~/.ssh/id_rsa")
@@ -226,3 +218,10 @@ def ssh():
     local("eval \"$(ssh-agent -s)\"")
     # Adds private key identities to the authentication agent
     local("ssh-add ~/.ssh/id_rsa")
+
+def start():
+    """Start the Sage Cell Server"""
+
+    sagecell_path = expanduser("~/sc_build/sage/sagecell")
+    print("Shut down the Sage Cell Server: Ctrl+c")
+    local("cd %s; ../sage web_server.py" % sagecell_path)
